Refactor: Navigator 컴포넌트 추상화
Changes to be committed:
modified:   src/ts/app/App.ts
modified:   src/ts/components/Navigator.ts
